+++ example/src/screens/Home/styles.ts
@@ -1,5 +1,9 @@
 import { createUseStyle } from '@themes';
 
-export default createUseStyle(() => ({
-  container: {},
+export default createUseStyle((t) => ({
+  container: {
+    backgroundColor: t.colors.secondary,
+    padding: 10,
+    borderWidth: 5,
+  },
 }));