+++ lib/src/types.d.ts
@@ -14,11 +14,13 @@ export type BaseStyleObj<BS extends BaseStyles<BS>> = BS extends NamedStyles<any
 export type CombinedStyleObj<S extends Styles<S>, BS extends BaseStyles<BS>, BSKey extends string> = BS extends undefined ? S : S & {
     [key in BSKey]: BS;
 };
-export type ThemeContextValue<T extends Themes, TP> = {
+export type ThemeContextValue<T extends Themes, ThemeKey extends string, TP> = {
     selectedTheme: ExtractThemeNames<T>;
     themes: T;
-    t: ExtractThemes<T>;
     themeParams: TP;
+    themeKey: ThemeKey;
+} & {
+    [key in ThemeKey]: ExtractThemes<T>;
 };
 export type ThemeBaseStylesContextValue<BS extends BaseStyles<BS>, BSKey extends string> = {
     baseStyles: BaseStyleObj<BS>;
@@ -28,7 +30,7 @@ export type ThemeDispatchContextValue<T extends Themes, TP> = {
     setTheme: (t: ExtractThemeNames<T>) => void;
     setThemeParams: (p: TP) => void;
 };
-export type ThemeContextProps<T extends Themes, BS extends Styles<BS>, BSKey extends string, TP> = {
+export type ThemeContextProps<T extends Themes, BS extends Styles<BS>, BSKey extends string, ThemeKey extends string, TP> = {
     initialThemeParams?: TP;
     themes: T | ((params: TP) => T);
     initialTheme: ExtractThemeNames<T>;
@@ -38,8 +40,9 @@ export type ThemeContextProps<T extends Themes, BS extends Styles<BS>, BSKey ext
     baseStylesKey?: BSKey;
     children: React.ReactNode;
     styleCacheManager?: StyleCacheManager<T, any, any>;
+    themeKey?: ThemeKey;
 };
-export type InitThemeProviderParams<T extends Themes, BS extends Styles<BS>, BSKey extends string, P = undefined> = Omit<ThemeContextProps<T, BS, BSKey, P>, 'children'>;
+export type InitThemeProviderParams<T extends Themes, BS extends Styles<BS>, BSKey extends string, ThemeKey extends string, P = undefined> = Omit<ThemeContextProps<T, BS, BSKey, ThemeKey, P>, 'children'>;
 export type StyleCreator<T extends Themes, S extends Styles<S>, P = undefined> = (theme: ExtractThemes<T>, params: P) => StyleObj<S>;
 export type BaseStyleCreator<T extends Themes, BS extends Styles<BS>> = (theme: ExtractThemes<T>) => StyleObj<BS>;
 export type StyleCacheManager<T extends Themes, S extends Styles<S>, P = undefined> = {